Gimme that stuff pls (#3)
* randomness libraries

+2D simplex noise lib
+quick seeded xor-shift RNG

* Fixed the SDL folder issue

Moved things around so SDL works again

* Added sdl2.dll to debug folder

This lets the exe run if you wanna do that

* fixed gitignore

* fixed gitignore again

* completely recreated project environement. previous file structure was disorganized

* added credits roll

some people still need to add their images

* cred done

* Deleted All of the Old

We needed to get rid of the old to make way for the new.

* Fix it Again

Again and Again

* Update Makefile

Made it work for Linux?

* Update Makefile

Finally have a fully working MakeFile for both Linux and Windows! Good work boys.

* setup lightweight cross-platform dev env in vs code

+added cross-platform, hands off makefile with powerful macros
+configured a vs code cross-platform build task and the build+debug launch.json
+hot key for debugging with G5
+notes: dependency for SDL on windows is self-contained, but linux must install thru their package manager

* fixed res artifacts from someone elses push to my repo

* clearing main

* trying to resolve conflicts

* readding gitignore and readme

* Delete Speedrun.exe

* Pull Request 6

github gets errors-- not merging errors-- ERRORS! amazing! Ill just replace all the files manually.

* Removed bin/include/.vscode

* base

Sets up the base code

* organization and movement

reorganizaed init, credits, and close functionality into their own functions to be called in main

added movement code based on cs1666 example into a rungame function to be called in main

* Added Block with Gravity and Movement

We added a block that moves based on key inputs and is affected by gravity. The block has a collider with the ground.

* got the main menu working

Networking team got the main menu working and tweaked the main loop some to account for it.

* Combined Menu, Credits, and Game

Got all three parts to run together

* Added Player Texture

Added Guy.png to the player rect

* Wiped for clarity

Wiped for clarity

* Hopefully this one works

This should upload all of the good stuff

* Ignore lib

* README update

Co-authored-by: Ryan Durkoske <rdurkoske@gmail.com>
Co-authored-by: alp203 <61202121+alp203@users.noreply.github.com>
Co-authored-by: lpc17 <63122782+lpc17@users.noreply.github.com>
Co-authored-by: Ruzakiff <ryanchenyang@gmail.com>
Co-authored-by: spap975 <48142512+spap975@users.noreply.github.com>
Co-authored-by: RobertXu09 <59585814+RobertXu09@users.noreply.github.com>
Co-authored-by: RobertXu09 <rox5@pitt.edu>
Co-authored-by: spap975 <spap975@hotmail.com>